Int32 == Int tests

+++ test/runtests.jl
@@ -2,11 +2,11 @@ using SupremumInfimum
 using Base.Test
 
 # write your own tests here
-@test leaf_type(Int) == Int64
+@test leaf_type(Int) == Int
 @test leaf_type(UInt8) == UInt8
 @test leaf_type(Real) == Float64
-@test leaf_type(Integer) == Int64
-@test leaf_type(Unsigned) == UInt64
+@test leaf_type(Integer) == Int
+@test leaf_type(Unsigned) == UInt
 
 @test sup(Set{Int}[]) == Set{Int}()
 @test inf(Float64[]) == Inf # Note the names match!
